pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "matrix-ai"
version = "0.1.0"
description = "AI service for Matrix EcoSystem"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "httpx==0.27.0",
    "pydantic==2.7.1",
    "python-json-logger==2.0.7",
    "cachetools==5.3.3",
    "huggingface-hub==0.23.0",
    "sentence-transformers==2.7.0",
    "faiss-cpu==1.8.0",
    "numpy==1.26.4",
    "orjson==3.10.3",
    "pyyaml==6.0.1",
    "tenacity==8.2.3",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]
